+++ src/olmo_core/data/__init__.py
@@ -1,11 +1,14 @@
 from .collator import DataCollator, PaddingDirection
 from .iterable_dataset import IterableDataset
 from .memmap_dataset import MemMapDataset, MemMapDatasetConfig, MemMapDType
+from .tokenizer import TokenizerConfig, TokenizerNames
 
 __all__ = [
     "MemMapDatasetConfig",
     "MemMapDataset",
     "MemMapDType",
+    "TokenizerConfig",
+    "TokenizerNames",
     "DataCollator",
     "PaddingDirection",
     "IterableDataset",